requirements.txt
python-3.11.4
# Core ML stack
torch==2.8.0
torchvision==0.23.0
ultralytics~=8.3.199

# ONNX inference
onnx>=1.14.0,<1.20.0
onnxruntime>=1.16.0

# Image processing
opencv-python-headless==4.10.0.84
pillow~=11.1.0
pillow-avif-plugin==1.5.2
piexif==1.1.3
imageio==2.37.0
scikit-image==0.25.2
albumentations==2.0.8
albucore==0.0.24

# Flask app
Flask==3.0.3
gunicorn==21.2.0
Werkzeug~=3.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
blinker==1.9.0

# Deployment + utilities
python-dotenv==1.1.1
requests~=2.31.0
psutil==7.0.0
filelock~=3.15.4
platformdirs~=4.2.2
setuptools~=69.0.2
wheel~=0.42.0

# Data + math
numpy~=1.26.4
scikit-learn==1.7.2
joblib==1.5.2
threadpoolctl==3.6.0
matplotlib==3.10.6
networkx==3.5
mpmath==1.3.0
python-dateutil==2.9.0.post0
future~=1.0.0

# Parsing + formatting
prettytable==3.16.0
pyparsing==3.2.3
PyYAML==6.0.2
typing_extensions~=4.12.2
decorator==5.1.1
easydict==1.13

# Security
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
pycparser==2.21

# Web + HTML
beautifulsoup4==4.13.5
soupsieve==2.5
charset-normalizer==3.3.2
certifi==2025.8.3
cachetools==5.5.2
colorama==0.4.6
tqdm==4.67.1
lazy_loader==0.4
Cython~=3.1.3
fonttools==4.59.2
kiwisolver==1.4.9
contourpy==1.3.3
brotlicffi==1.0.9.2
anyio==4.7.0
attrs~=25.3.0
Pygments~=2.19.1
pydantic==2.11.9
pydantic_core==2.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
tifffile==2025.9.20